Patch bug in sys2

+++ src/AdamsIntegrator.hpp
@@ -162,7 +162,7 @@ void Adams_BM<T>::step(int n) {
     this->sys(tmp2, tmp, this->t[n+1]);
     tmp *= b_AM[0];
     // Now tmp goes back to being an aggregator
-    for (size_t i = 1; i < order; i++) {
+    for (int i = 1; i < order; i++) {
         this->sys(this->y[n-i+1], ydot, this->t[n-i+1]);
         ydot *= b_AM[i];
         tmp += ydot;
@@ -193,7 +193,7 @@ void Adams_BM<T>::iterate(double t_initial, double t_final) {
     // Set up the t grid
     this->t[0] = t_initial;
 
-    for (size_t n=1; n<npoints; n++){
+    for (int n=1; n<npoints; n++){
         this->t[n] = this->t[n-1] + this->dt;
     }
     this->run_steps();
@@ -205,12 +205,12 @@ void Adams_BM<T>::run_steps(){
     assert(this->t.size() >= order);
 
     // initialise enough points for multistepping to get going
-    for (size_t n = 0; n < order; n++) {
+    for (int n = 0; n < order; n++) {
         this->step_rk4(n);
     }
     // Run those steps
     std::cout << "[ sim ]                       ";
-    for (size_t n = this->order; n < this->t.size()-1; n++) {
+    for (int n = this->order; n < this->t.size()-1; n++) {
         std::cout << "\r[ sim ] t="
                   << std::left<<std::setfill(' ')<<std::setw(6)
                   << this->t[n] << std::flush;
